+++ internal/http/handlers/self_account.go
@@ -76,7 +76,7 @@ type selfAccountSettingsRequest struct {
 }
 
 type selfAccountSettingsResponse struct {
-	AutoTopupFailures      int32      `json:"auto_topup_failures"`
+	AutoTopupFailures      int64      `json:"auto_topup_failures"`
 	Currency               string     `json:"currency"`
 	LowBalanceThreshold    *int64     `json:"low_balance_threshold,omitempty"`
 	AutoTopupEnabled       bool       `json:"auto_topup_enabled"`
